refactor: 공통 요소 컴포넌트로 분리
